ci: less clutter in the timeout test
