+++ detox/test/package.json
@@ -12,15 +12,15 @@
     "detox-server": "detox run-server",
     "e2e:ios": "detox test --configuration ios.sim.release --debug-synchronization",
     "e2e:jest-circus:ios": "detox test --configuration ios.sim.release --debug-synchronization -o e2e/config-circus.json",
-    "e2e:jest-circus-timeout:ios": "node scripts/assert_timeout.js npm run e2e:jest-circus:ios -- --take-screenshots failing e2e/01.sanity.test.js",
+    "e2e:jest-circus-timeout:ios": "node scripts/assert_timeout.js npm run e2e:jest-circus:ios -- --take-screenshots failing e2e/26.timeout.test.js",
     "e2e:ios-ci": "detox test --configuration ios.sim.release --workers 3 --debug-synchronization --take-screenshots all --record-logs all --record-timeline all -- --coverage",
-    "e2e:ios-timeout-ci": "node scripts/assert_timeout.js npm run e2e:ios-ci -- e2e/01.sanity.test.js e2e/02.matchers.test.js e2e/03.actions.test.js",
+    "e2e:ios-timeout-ci": "node scripts/assert_timeout.js npm run e2e:ios-ci -- e2e/26.timeout.test.js",
     "e2e:android": "detox test --configuration android.emu.release",
     "e2e:jest-circus:android": "detox test --configuration android.emu.release -o e2e/config-circus.json",
-    "e2e:jest-circus-timeout:android": "node scripts/assert_timeout.js npm run e2e:jest-circus:android -- --take-screenshots failing e2e/01.sanity.test.js",
+    "e2e:jest-circus-timeout:android": "node scripts/assert_timeout.js npm run e2e:jest-circus:android -- --take-screenshots failing e2e/26.timeout.test.js",
     "e2e:android-debug": "detox test --configuration android.emu.debug",
     "e2e:android-ci": "detox test --configuration android.emu.release --workers 3 --take-screenshots all --record-logs all --record-timeline all --headless --loglevel verbose --jest-report-specs -- --coverage",
-    "e2e:android-timeout-ci": "node scripts/assert_timeout.js npm run e2e:android-ci -- e2e/01.sanity.test.js e2e/02.matchers.test.js e2e/03.actions.test.js",
+    "e2e:android-timeout-ci": "node scripts/assert_timeout.js npm run e2e:android-ci -- e2e/26.timeout.test.js",
     "build:ios": "detox build --configuration ios.sim.release",
     "build:android": "detox build --configuration android.emu.release",
     "build:android-debug": "detox build --configuration android.emu.debug",
